fix!: rename translate2 task to translate and delete old translation task
