+++ .gitignore
@@ -36,3 +36,8 @@ site/
 examples/*/.cache/
 examples/*/input/
 examples/*/output/
+
+*.log
+*.out
+*.err
+*.txt
