+++ src/tigerflow_ml/text/translate/detection.py
