--- src/tigerflow_ml/text/translate2/__init__.py
@@ -1,3 +1,0 @@
-from tigerflow_ml.text.translate2.slurm import Translate
-
-__all__ = ["Translate"]
